Refactored Flask server out of instantiator. Added 'constants' file for easily modifying manager-specific variables.
