+++ workshop-manager/bin/RequestHandler/client_updater.py
@@ -1,11 +1,11 @@
-import webdata_aggregator, time
+import time
+from DataAggregation.webdata_aggregator import getAvailableWorkshops
 from socketio import socketio_manage
 from socketio.namespace import BaseNamespace
 from socketio.mixins import BroadcastMixin
 
 
-class SocketIOApp(object):
-    """Stream sine values"""
+class RequestHandlerApp(object):
     def __call__(self, environ, start_response):
         if environ['PATH_INFO'].startswith('/socket.io'):
             socketio_manage(environ, {'': QueueStatusHandler})
@@ -13,7 +13,7 @@ class SocketIOApp(object):
 
 class QueueStatusHandler(BaseNamespace, BroadcastMixin):
     def on_run(self):
-        workshops = webdata_aggregator.getAvailableWorkshops()
+        workshops = getAvailableWorkshops()
         sizes = []
         for w in workshops:
             tmp = [w.workshopName, w.q.qsize()]
@@ -21,7 +21,7 @@ class QueueStatusHandler(BaseNamespace, BroadcastMixin):
             self.emit('sizes', tmp)
 
         while True:
-            curr_workshops = webdata_aggregator.getAvailableWorkshops()
+            curr_workshops = getAvailableWorkshops()
             for w in curr_workshops:
                 wq = filter(lambda x: x[0] == w.workshopName, sizes)[0]
