+++ workshop-manager/bin/WebServer/static/connect.PNG
