+++ workshop-manager/bin/WebServer/static/connected.PNG
