+++ workshop-manager/bin/WebServer/static/js/socket.io.js
