+++ workshop-manager/bin/WebServer/static/noise-diagonal.png
