+++ workshop-manager/bin/instantiator.py
@@ -1,141 +1,10 @@
-# Project imports
-import DataAggregation.webdata_aggregator, VMStateManager.vbox_monitor, os, time, logging, signal
-
-# gevent imports
+import DataAggregation.webdata_aggregator, VMStateManager.vbox_monitor, logging, signal
+import WebServer
 import gevent, gevent.monkey; gevent.monkey.patch_all()
-import threading
 from gevent.pywsgi import WSGIServer
-from gevent.lock import BoundedSemaphore
-
-# Flask imports
-from flask import Flask, make_response, render_template, send_from_directory, jsonify
-
-# nocache imports
-from functools import wraps, update_wrapper
-
 from socketio.server import SocketIOServer
-from DataAggregation.client_updater import SocketIOApp
-from DataAggregation.webdata_aggregator import zip_files
-
-# Webserver commands
-app = Flask(__name__)
-app.debug = True
-
-threadsToRun = []
-threadsToRunSem = threading.Semaphore(1)
-threadTime = 2
-threadHandlerSem = threading.Semaphore(1)
-
-zipSem = BoundedSemaphore(1)
-zipClearTime = 15
-i = 0
-
-
-def clearZip(zip_file_name):
-    time.sleep(zipClearTime)
-    os.remove(zip_file_name)
-
-
-def threadHandler():
-    while True:
-        threadHandlerSem.acquire()
-        if len(threadsToRun):
-            for thread in threadsToRun:
-                t = thread
-                threadsToRun.remove(thread)
-                t.start()
-        threadHandlerSem.release()
-        time.sleep(threadTime)
-
-
-def nocache(view):
-    @wraps(view)
-    def no_cache(*args, **kwargs):
-        response = make_response(view(*args, **kwargs))
-        #response.headers['Last-Modified'] = datetime.now()
-        response.headers['Cache-Control'] = 'public, no-store, no-cache, must-revalidate, post-check=0, pre-check=0, max-age=0'
-        response.headers['Pragma'] = 'no-cache'
-        response.headers['Expires'] = '0'
-        return response
-    return update_wrapper(no_cache, view)
-
-
-@app.route('/WorkshopData/<path:filename>', methods=['GET', 'POST'])
-@nocache
-def download(filename):
-    downloads = os.path.join(app.root_path, "WorkshopData/")
-    return send_from_directory(directory=downloads, as_attachment=True, filename=filename, mimetype='application/octet-stream')
-
-
-# Simple catch-all server
-@app.route('/', defaults={'path': ''}, methods=['GET', 'POST'])
-@nocache
-def catch_all(path):
-    """ Handles all requests to the main index page of the Web Server. """
-    return render_template('index.html', workshops=DataAggregation.webdata_aggregator.getAvailableWorkshops())
-
-
-@app.route('/checkout/ms-rdp/<workshopName>')
-def checkoutRDP(workshopName):
-    global i
-    workshop = filter(lambda x: x.workshopName == workshopName, DataAggregation.webdata_aggregator.getAvailableWorkshops())[0]
-    if workshop.q.qsize():
-        unit = workshop.q.get()
-        threadsToRunSem.acquire()
-        DataAggregation.webdata_aggregator.putOnHold(unit)
-        threadsToRun.append(threading.Thread(target=DataAggregation.webdata_aggregator.checkoutUnit, args=(unit,)))
-        threadsToRunSem.release()
-        rdpPaths = unit.rdp_files
-
-        if len(rdpPaths) is 1:
-            return render_template('download.html', download_path=rdpPaths[0], download_type="windows")
-
-        zipSem.acquire()
-        zip_file_name = "Workshop" + str(i) + ".zip"
-        zip_file_path = os.path.join("WorkshopData", workshopName, zip_file_name)
-        i += 1
-        zip_files(rdpPaths, zip_file_path)
-        threadsToRun.append(threading.Thread(target=clearZip, args=(zip_file_path,)))
-        zipSem.release()
-        return render_template('download.html', download_path=zip_file_path, download_type="windows")
-    else:
-        return "Sorry, there are no workshops available."
-
-
-@app.route('/checkout/rdesktop/<workshopName>')
-def checkoutrdesktop(workshopName):
-    global i
-    workshop = filter(lambda x: x.workshopName == workshopName, DataAggregation.webdata_aggregator.getAvailableWorkshops())[0]
-    if workshop.q.qsize():
-        unit = workshop.q.get()
-        threadsToRunSem.acquire()
-        DataAggregation.webdata_aggregator.putOnHold(unit)
-        threadsToRun.append(threading.Thread(target=DataAggregation.webdata_aggregator.checkoutUnit, args=(unit,)))
-        threadsToRunSem.release()
-        rdesktopPaths = unit.rdesktop_files
-
-        if len(rdesktopPaths) is 1:
-            return render_template('download.html', download_path=rdesktopPaths[0], download_type="linux")
-
-        zipSem.acquire()
-        zip_file_name = unit.workshopName + "_" + str(i) + ".zip"
-        zip_file_path = os.path.join("WorkshopData", workshopName, zip_file_name)
-        i += 1
-        zip_files(rdesktopPaths, zip_file_path)
-        threadsToRun.append(threading.Thread(target=clearZip, args=(zip_file_path,)))
-        zipSem.release()
-        return render_template('download.html', download_path=zip_file_path, download_type="linux")
-    else:
-        return "Sorry, there are no workshops available."
-
-@app.route('/getQueueSize/<workshopName>')
-def giveQueueSize(workshopName):
-    """ Catch AJAX Requests for Queue Size. """
-    availableWorkshops = DataAggregation.webdata_aggregator.getAvailableWorkshops()
-    if (availableWorkshops):
-        return jsonify(filter(lambda x: x.workshopName == workshopName, availableWorkshops)[0].q.qsize())
-    else:
-        return jsonify("0")
+from RequestHandler.client_updater import RequestHandlerApp
+from WebServer.flask_server import app
 
 
 def signal_handler(signal, frame):
@@ -161,13 +30,13 @@ if __name__ == '__main__':
     logging.basicConfig(level=logging.DEBUG)
 
     httpServer = WSGIServer(('0.0.0.0', 8080), app)
-    sio_server = SocketIOServer(('0.0.0.0', 9090), SocketIOApp(), namespace="socket.io")
+    sio_server = SocketIOServer(('0.0.0.0', 9090), RequestHandlerApp(), namespace="socket.io")
     stateAssignmentThread = gevent.spawn(VMStateManager.vbox_monitor.manageStates)
     restoreThread = gevent.spawn(VMStateManager.vbox_monitor.makeRestoreToAvailableState)
     srvGreenlet = gevent.spawn(httpServer.start)
     ioGreenlet = gevent.spawn(sio_server.serve_forever)
     dataAggregator = gevent.spawn(DataAggregation.webdata_aggregator.aggregateData)
-    threadHandler = gevent.spawn(threadHandler)
+    threadHandler = gevent.spawn(WebServer.flask_server.threadHandler)
 
     try:
         # Let threads run until signal is caught
